requirements.txt
edge-tts                # 文本转语音服务
torch==2.3              # PyTorch深度学习框架
torchaudio              # PyTorch的音频处理库
funasr>=1.1.3           # 语音识别框架
fastapi>=0.111.1        # 快速构建API的框架
uvicorn                 # ASGI服务器，用于运行FastAPI应用
python-dotenv           # 从.env文件加载环境变量
transformers==4.51.0    # 自然语言处理模型库
FlagEmbedding==1.2.3    # 分块嵌入库
pydantic>=2.0.0         # 数据模型验证库
mysql-connector-python>=8.1.0  # MySQL数据库连接器
FastMCP                 # MPC通信协议库
dingtalk_stream         # 钉钉机器人SDK
alibabacloud_dingtalk   # 钉钉机器人API客户端库
numpy>=1.21.0           # 数值计算库
sentence-transformers>=2.2.0  # 文本嵌入库
requests>=2.25.0        # HTTP请求库
PyPDF2>=1.27.0          # PDF处理库
python-docx>=0.8.11     # 读取Word文档库
chardet>=4.0.0          # 字符编码检测库
chromadb>=0.2.0         # 基于SQLite的向量数据库
itsdangerous>=2.1.2     # 安全相关的工具库
matplotlib              # 绘图库
omegaconf               # 用于配置文件的库
tqdm                    # 用于显示训练进度的库
pyyaml                  # 用于解析YAML配置文件
deepspeed               # 用于加速训练的库
sentencepiece           # 用于处理文本的分词库
einops                  # 张量操作库
scipy                   # 科学计算库
soundfile               # 音频文件处理库
packaging               # 用于处理Python包的元数据
pandas                  # 数据处理库
huggingface_hub
WeTextProcessing; platform_machine != "Darwin"
wetext; platform_system == "Darwin"

# sql_agent 依赖
langchain>=0.1.0
langchain-openai>=0.0.5
langgraph>=0.0.40
langchain-community>=0.0.10
sqlalchemy>=2.0.0
sqlparse>=0.4.4
termcolor>=1.1.0

# any4dh 数字人依赖
opencv-python-headless      # 计算机视觉库 (无GUI版本)
python_speech_features      # 语音特征提取
numba                       # 数值计算加速
resampy                     # 音频重采样
aiortc                      # WebRTC 实时通信
aiohttp                     # 异步HTTP客户端
av                          # 音视频处理库
face_alignment              # 人脸对齐库
imageio-ffmpeg              # FFmpeg 封装
beautifulsoup4              # HTML解析库